+++ es-core/src/components/ComponentList.cpp
@@ -177,7 +177,7 @@ void ComponentList::render(const Eigen::Affine3f& parentTrans)
 	bool drawAll;
 	for(unsigned int i = 0; i < mEntries.size(); i++) {
 		auto& entry = mEntries.at(i);
-		drawAll = !mFocused || i != mCursor;
+		drawAll = !mFocused || (int)i != mCursor;
 		for(auto it = entry.data.elements.begin(); it != entry.data.elements.end(); it++) {
 			if(drawAll || it->invert_when_selected) {
 				it->component->render(trans);
